+++ neoforge/src/main/java/com/jahirtrap/walljump/init/mixin/PlayerMixin.java
@@ -18,8 +18,8 @@ public abstract class PlayerMixin {
     @ModifyArg(method = "causeFallDamage", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/LivingEntity;causeFallDamage(FFLnet/minecraft/world/damagesource/DamageSource;)Z"), index = 0)
     private float causeFallDamage(float value) {
         if (value > 3 && value <= ServerConfig.minFallDistance) {
-            playSound(SoundEvents.GENERIC_SMALL_FALL, 0.5F, 1.0F);
-            return 3.0F;
+            playSound(SoundEvents.GENERIC_SMALL_FALL, 0.5f, 1f);
+            return 3;
         }
 
         return value;