+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Some fixes
+- Reimplemented falling sound (<a href="https://github.com/jahirxtrap/walljump/issues/24">Issue #24</a>)